Usuarios ya no pueden crear pqrs con otro usuarioID
Tambien se cambio la id de rol de Long a Integer y se elimino Estado de la tabla de las pqrs para usar solamente Id Estado, En otro commit se añadira la tabla estados y el json de pqrs incluira el json de estados.

+++ src/main/java/com/feature5/pqrs/DTO/RolDTO.java
@@ -1,24 +1,24 @@
 package com.feature5.pqrs.DTO;
 
 public class RolDTO {
-    private Long idRol;
+    private Integer idRol;
     private String descripcion;
 
     // Constructor por defecto
     public RolDTO() {}
 
     // Constructor con parámetros
-    public RolDTO(Long idRol, String descripcion) {
+    public RolDTO(Integer idRol, String descripcion) {
         this.idRol = idRol;
         this.descripcion = descripcion;
     }
 
     // Getters y setters
-    public Long getIdRol() {
+    public Integer getIdRol() {
         return idRol;
     }
 
-    public void setIdRol(Long idRol) {
+    public void setIdRol(Integer idRol) {
         this.idRol = idRol;
     }
